Add timetable viewing & submission
- split JS & CSS to load required files only
- button redesigns
- added timetable buttons to users & dozenten in users list view
- split viewing & editing timetables in different blade files
    - editing is only possible when viewing own timetable
    - admins can see the timetable for each dozent, everyone else only for a studiengang
- fix missing styling of error messages
- change StundenController.php index() to be able to show timetables for any dozent
- automatically change the status of the dozent between Nicht abgegeben & In Arbeit when the amount of stunden is >0 or =0
- added functionality to submit the timetable (set plan_abgegeben to 1)
- translate profile page into german
- fix calculateEndTime() not calculating the correct end_time for courses starting at 11:45 that are longer than 4 sws
