+++ vite.config.js
@@ -8,6 +8,9 @@ export default defineConfig({
                 'resources/css/app.css',
                 'resources/js/app.js',
                 'resources/js/timetable_crud.js',
+                'resources/css/timetable_crud.css',
+                'resources/js/timetable_show.js',
+                'resources/css/timetable_show.css',
             ],
             refresh: true,
         }),
